% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTeamStats2023.R
\name{calculateTeamStats2023}
\alias{calculateTeamStats2023}
\title{This function calculates the statistics for a specific team based on the games played in 2023.
It calculates the total number of games played, total wins, total losses, and winning percentage.}
\usage{
calculateTeamStats2023(data, teamName)
}
\arguments{
\item{data}{A tibble containing game data, including columns for 'Home_Team', 'Away_Team', 'Home_Score', and 'Away_Score'.}

\item{teamName}{The name of the team for which statistics are to be calculated.}
}
\value{
A tibble summarizing the team statistics for 2023, including total games played, total wins, total losses, and winning percentage.
}
\description{
This function calculates the statistics for a specific team based on the games played in 2023.
It calculates the total number of games played, total wins, total losses, and winning percentage.
}
\examples{
calculateTeamStats2023(games_2023, "SUU")

}
