% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbbga23}
\alias{cbbga23}
\title{College Basketball 2023 Data Set}
\format{
A data frame with 6 variables: \code{Date_of_Game}, \code{Away_Team},
\code{Away_Score}, \code{Home_Team}, \code{Home_Score}, and \code{Score_Difference}.
}
\usage{
cbbga23
}
\description{
This data set contains information on college basketball games from the
2023 season. Information includes the date of the game played, the names
and scores for the home and away teams, and the score difference for each game.
}
\keyword{datasets}
