% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filteredTeamGames.R
\name{filteredTeamGames}
\alias{filteredTeamGames}
\title{This function filters the input tibble to only include games played by the
specified team, whether as a home team or away team.}
\usage{
filteredTeamGames(data, team)
}
\arguments{
\item{data}{The input tibble containing game data.}

\item{team}{The name of the team to filter for.}
}
\value{
A tibble containing only the games played by the specified team.
}
\description{
This function filters the input tibble to only include games played by the
specified team, whether as a home team or away team.
}
\examples{
filtered_games <- filteredTeamGames(games_2023, "SUU")

}
