% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGamesByTeam.R, R/plotWinsLosses.R
\name{plotGamesByTeam}
\alias{plotGamesByTeam}
\title{This function creates a bar plot showing the winning percentage of each team.
It takes a tibble containing team statistics as input and plots the winning percentage for each team.}
\usage{
plotGamesByTeam(data)

plotGamesByTeam(data)
}
\arguments{
\item{data}{A tibble containing team statistics, including columns for 'teamName' and 'winningPercentage'.}
}
\value{
A ggplot object representing a bar plot showing the winning percentage of each team.

A ggplot object representing a bar plot showing the winning percentage of each team.
}
\description{
This function creates a bar plot showing the winning percentage of each team.
It takes a tibble containing team statistics as input and plots the winning percentage for each team.

This function creates a bar plot showing the winning percentage of each team.
It takes a tibble containing team statistics as input and plots the winning percentage for each team.
}
\examples{
plotGamesByTeam(team_data)

plotGamesByTeam(team_data)

}
